+++ frontend/app/components/container/ItemsCard/ItemsCard.hooks.ts
